backend_config/requirements.txt
django==4.1.5
djangorestframework==3.12.0
mysqlclient==2.1.1
pika==1.3.1
django-mysql==4.8.0
django-cors-headers==3.13.0